(REF) Convert TranslatorTest to MockTranslatorTest with many small files
