+++ ext/afform/mock/ang/translate/tEmptyDiv.aff.html
@@ -1,0 +1,1 @@
+<div></div>
