+++ ext/afform/mock/ang/translate/tEmptyDiv.rendered.html
@@ -1,0 +1,1 @@
+<div></div>
