+++ ext/afform/mock/ang/translate/tEmptyDiv.strings.php
@@ -1,0 +1,1 @@
+<?php return [];
