+++ ext/afform/mock/ang/translate/tEmptyDiv.translate.php
@@ -1,0 +1,1 @@
+<?php return [];
