+++ ext/afform/mock/ang/translate/tOriginalRich.aff.html
@@ -1,0 +1,1 @@
+<af-form ctrl="afform"><p class="af-text"><em>Localizable</em> Text</p></af-form>
