+++ ext/afform/mock/ang/translate/tOriginalRich.rendered.html
@@ -1,0 +1,1 @@
+<af-form ctrl="afform" ng-form="tOriginalRich"><p class="af-text"><em>Localizable</em> Text</p></af-form>
