+++ ext/afform/mock/ang/translate/tOriginalRich.translate.php
@@ -1,0 +1,1 @@
+<?php return [];
