+++ ext/afform/mock/ang/translate/tStaticText.rendered.html
@@ -1,0 +1,1 @@
+<div><span>Static Text</span></div>
