+++ ext/afform/mock/ang/translate/tStaticText.strings.php
@@ -1,0 +1,1 @@
+<?php return [];
