+++ ext/afform/mock/ang/translate/tStaticText.translate.php
@@ -1,0 +1,3 @@
+<?php return [
+  "Static Text" => "This should not be replaced because it was just static text.",
+];
