+++ ext/afform/mock/ang/translate/tTranslatedBr1.strings.php
@@ -1,0 +1,3 @@
+<?php return [
+  "Localizable<br> Text",
+];
