+++ ext/afform/mock/ang/translate/tTranslatedBr2.aff.html
@@ -1,0 +1,1 @@
+<af-form ctrl="afform"><p class="af-text">Localizable<br/> Text</p></af-form>
