+++ ext/afform/mock/ang/translate/tTranslatedBr2.rendered.html
@@ -1,0 +1,1 @@
+<af-form ctrl="afform" ng-form="tTranslatedBr2"><p class="af-text">Texte<br> localisable</p></af-form>
