+++ ext/afform/mock/ang/translate/tTranslatedBr2.translate.php
@@ -1,0 +1,3 @@
+<?php return [
+  "Localizable<br> Text" => "Texte<br> localisable",
+];
