+++ ext/afform/mock/ang/translate/tTranslatedDefn.rendered.html
@@ -1,0 +1,11 @@
+<af-form ctrl="afform" ng-form="tTranslatedDefn">
+  <af-entity data="{source: 'Hello world 2'}" type="Individual" name="Individual1" label="Individual 1" actions="{create: true, update: true}" security="FBAC" autofill="user"></af-entity>
+  <fieldset af-fieldset="Individual1" class="af-container" af-title="Individual 1">
+    <div class="af-container">
+      <div class="af-container af-layout-inline">
+        <af-field name="first_name" defn="{input_attrs: {placeholder: '(Your given name)', size: '30', label: 'Pr\u00e9nom'}, label: 'Given Name', required: false, name: 'first_name', data_type: 'String', fk_entity: null, dfk_entities: null, options: false, input_type: 'Text', help_pre: null, help_post: null, search_range: false}"></af-field>
+      </div>
+    </div>
+  </fieldset>
+  <button class="af-button btn btn-primary" crm-icon="fa-check" ng-click="afform.submit()">Envoyer</button>
+</af-form>
