+++ ext/afform/mock/ang/translate/tTranslatedDefn.strings.php
@@ -1,0 +1,6 @@
+<?php return [
+  '(Your given name)',
+  'Given Name',
+  'Individual 1',
+  'Submit',
+];
