+++ ext/afform/mock/ang/translate/tTranslatedDefn.translate.php
@@ -1,0 +1,1 @@
+<?php return [];
