+++ ext/afform/mock/ang/translate/tTranslatedEm.rendered.html
@@ -1,0 +1,1 @@
+<af-form ctrl="afform" ng-form="tTranslatedEm"><p class="af-text">Texte <em>localisable</em></p></af-form>
