+++ ext/afform/mock/ang/translate/tTranslatedEm.strings.php
@@ -1,0 +1,3 @@
+<?php return [
+  "<em>Localizable</em> Text",
+];
