+++ ext/afform/mock/ang/translate/tTranslatedEm.translate.php
@@ -1,0 +1,3 @@
+<?php return [
+  "<em>Localizable</em> Text" => "Texte <em>localisable</em>",
+];
