+++ ext/afform/mock/ang/translate/tTranslatedMixed.aff.html
@@ -1,0 +1,3 @@
+<af-form ctrl="afform"><span>Static Start</span>
+<p class="af-text">Localizable Text</p>
+<span>Static End</span></af-form>
