+++ ext/afform/mock/ang/translate/tTranslatedMixed.rendered.html
@@ -1,0 +1,3 @@
+<af-form ctrl="afform" ng-form="tTranslatedMixed"><span>Static Start</span>
+<p class="af-text">Texte localisable</p>
+<span>Static End</span></af-form>
