+++ ext/afform/mock/ang/translate/tTranslatedMixed.strings.php
@@ -1,0 +1,3 @@
+<?php return [
+  "Localizable Text",
+];
