+++ ext/afform/mock/ang/translate/tTranslatedMixed.translate.php
@@ -1,0 +1,5 @@
+<?php return [
+  "Localizable Text" => "Texte localisable",
+  "Static Start" => "Fake news!",
+  "Static End" => "Fake news!",
+];
