+++ ext/afform/mock/ang/translate/tUntranslated1.aff.html
@@ -1,0 +1,1 @@
+<af-form ctrl="afform"><p class="af-text">Localizable Text</p></af-form>
