+++ ext/afform/mock/ang/translate/tUntranslated1.rendered.html
@@ -1,0 +1,1 @@
+<af-form ctrl="afform" ng-form="tUntranslated1"><p class="af-text">Localizable Text</p></af-form>
