+++ ext/afform/mock/ang/translate/tUntranslated1.strings.php
@@ -1,0 +1,3 @@
+<?php return [
+  "Localizable Text",
+];
