+++ ext/afform/mock/ang/translate/tUntranslated1.translate.php
@@ -1,0 +1,1 @@
+<?php return [];
