+++ ext/afform/mock/ang/translate/tUntranslated3.rendered.html
@@ -1,0 +1,3 @@
+<af-form ctrl="afform" ng-form="tUntranslated3"><p class="af-text">Localizable<br> Text</p>
+<div class="af-markup">Boo<br>Boo</div>
+<button>Go time!</button></af-form>
