+++ ext/afform/mock/ang/translate/tUntranslated3.strings.php
@@ -1,0 +1,5 @@
+<?php return [
+  "Localizable<br> Text",
+  "Boo<br>Boo",
+  "Go time!",
+];
