+++ ext/afform/mock/ang/translate/tUntranslated3.translate.php
@@ -1,0 +1,1 @@
+<?php return [];
